requirements.txt
openai
pydantic
requests
streamlit
python-dotenv